+++ src/cli.ts
@@ -37,6 +37,7 @@ import { designmdCommand } from "./commands/designmd.js";
 import { schemaCommand } from "./commands/schema.js";
 import { memoryCommand } from "./commands/memory.js";
 import { changelogCommand } from "./commands/changelog.js";
+import { a11yLintCommand } from "./commands/a11y-lint.js";
 import { ingestFigmaDsCommand } from "./commands/ingest-figma-ds.js";
 import { synthesizeConventionsCommand } from "./commands/synthesize-conventions.js";
 
@@ -77,6 +78,7 @@ COMMANDS[designmdCommand.name] = designmdCommand;
 COMMANDS[schemaCommand.name] = schemaCommand;
 COMMANDS[memoryCommand.name] = memoryCommand;
 COMMANDS[changelogCommand.name] = changelogCommand;
+COMMANDS[a11yLintCommand.name] = a11yLintCommand;
 COMMANDS[ingestFigmaDsCommand.name] = ingestFigmaDsCommand;
 COMMANDS[synthesizeConventionsCommand.name] = synthesizeConventionsCommand;